soc: microchip: mpfs: add auto-update subdev to system controller
The PolarFire SoC's system controller offers In-Application Programming
(IAP) functionality. IAP enables the FPGA to reprogram itelf & will be
implemented as an FPGA Manager. The specific mode that will be supported
by the FPGA Manger subsystem is "auto-update". In this mode, a bitstream
is programmed to the SPI flash but will not be loaded until the device
is reset.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>