mailbox: mpfs: read the system controller's status
Some services explicitly return an error code in their response, but
others rely on the system controller to set a status in its status
register. The meaning of the bits varies based on what service is
requested, so pass it back up to the driver that that requested the
service in the first place. The field in the message struct already
existed, but was unused until now.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>