dt-bindings: riscv: update microchip.yaml's maintainership
Daire and I are the platform maintainers for Microchip's RISC-V
FPGAs. Update the maintainers in microchip.yaml to reflect this and
explicitly add the binding to the SoC's MAINTAINERS entry.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
---
Two patches seemed overkill for this, but scream and I will split them.
I figured I would take this for 6.1 myself on top of my other changes
to microchip.yaml.