riscv: replace SOC_MICROCHIP_POLARFIRE with ARCH_MICROCHIP
Per discussion at LPC, the SoC/platform Kconfig symbols in RISC-V land
are to be converted to ARCH_FOO from SOC_FOO so that the symbols are
consistent between the "incumbents" & those coming from ARM64 etc.

While we are already doing a rename, drop the "POLARFIRE" from the name
of the symbol for consistency's sake.

The SOC_MICROCHIP_POLARFIRE symbol is kept for one release to ease the
transition.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>