rtc: mpfs: replace SOC_MICROCHIP_POLARFIRE with ARCH_MICROCHIP
At LPC, we decided that for the sake of consistency the SOC_FOO symbols
in arch/riscv should be replaced with ARCH_FOO so that we could have a
consistent policy regardless of whether a vendor had ARM legacy or was
a RISC-V "incumbent".

As the PolarFire SoC RTC depends on the SOC_ symbol convert it, and
while we are at it, add the ability to compile test it for the craic.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>