riscv: replace SOC_SIFIVE with ARCH_SIFIVE
Per discussion at LPC, the SoC/platform Kconfig symbols in RISC-V land
are to be converted to ARCH_FOO from SOC_FOO so that the symbols are
consistent between the "incumbents" & those coming from ARM64 etc.

Predictably, SOC_SIFIVE becomes ARCH_SIFIVE. The ARCH_SIFIVE symbol is
kept for a release to ease the transition.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>