riscv: convert SOC_CANAAN to ARCH_CANAAN
As per the discussions at LPC, the SOC_ symbols in Kconfig.socs are to
be converted to ARCH_ for consistency between "incumbent" and incoming
vendors who have a legacy in other architectures. SOC_CANAAN therefore
becomes ARCH_CANAAN.

The SOC_CANAAN symbol is kept for a release to ease the transition.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>