RISC-V, dt-bindings: add riscv,isa-base property
I'm not entirely sure if this is needed, but I felt we still needed a
mechanism for communicating the base ISA. Perhaps the i here should not
even be present, but a way to encode the bit-width is missing from my
key-value stuff. Very much open to suggestions on this aspect.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>