media: mchp_dcsmi: permit changing h264 video resolution while streaming
The H.264 encoding section of the DSCMI pipeline supports changing the
resolution while the pipeline is streaming. The current webdemo for the
DSCMI IP uses devmem to change the resolution of the video stream in
response to commands from the user, rather than going through the
driver.
When IO_STRICT_DEVMEM is enabled, which currently is in mpfs_defconfig,
using devmem to interact with memory that has been bound to a driver is
not possible, causing the demo to malfunction. Besides, using devmem to
poke at hardware with a bound driver is extremely fragile to begin with.

Add the ability to change the resolution into the driver, so that the
webdemo functions as expected with our default configs.

Signed-off-by: Shravan Chippa <shravan.chippa@microchip.com>
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>