spi: microchip-core-qspi: fix transfer_one() for reads
If a transfer is a pure read, only read_op() is called by the driver,
which only reads from the rx FIFO, without actually affecting the bus.
Without affecting the bus, nothing will have actually populated the
FIFOs and the driver will spin forever, waiting for data that will not
arrive. Use write_read_op() instead, modifying it so that it will emit
dummy data to generate clock pulses if the tx buffer is empty, thereby
populating the rx FIFO with the requested read.

Fixes: 858175ed5f6c ("spi: microchip-core-qspi: Add regular transfers")
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>