ARM: configs: at91: sam9x75 Add switch support
The sam9x75 curiosity has a connector allowing it to support daughter cards
containing both single-port ethernet PHYs and ethernet switches. When an
ethernet switch is attached, the necessary DSA drivers and packages need to be
included in the kernel. This patch allows the SAM9x75 board to support the
KSZ8795, KSZ8895, KSZ9477, KSZ9563, and LAN9646 EDS2 boards.

Signed-off-by: Jerry Ray <jerry.ray@microchip.com>
Reviewed-by: Hari Prasath Gujulan Elango <hari.prasathge@microchip.com>