clk: at91: sama7d65: add missing GPU peripheral and generic clocks
The GPU peripheral and generic clocks are both required to use the
Vivante GC520UL GPU embedded into SAMA7D65 SoCs.

Please note that the `max` field for the `gpu_gclk` entry in the
`sama7d65_gck[]` array is set on purpose to 533000001 rather than
533000000 Hz.

Indeed, if set to 533000000 instead, when the GPU GCK clock changes the
clock rate of its parent, the GPU PLL, this clock rate is actually set
to 533000001. Then, when the actual clock rate of the GPU GCK is
computed from the 533000001 Hz of the GPU PLL, the final value is
also 533000001 for the GPU GCK.

However, with a max value of 533000000, the operation fails with a -22
(-EINVAL) error code. So, setting this max value at 533000001 fixes this
issue.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Reviewed-by: Ryan Wanner <ryan.wanner@microchip.com>