riscv: dts: microchip: use ihc node to use ipc driver for beaglev-fire
The BeagleV-Fire now supports the new MI-V IHC IP v2 available in the
Libero IP catalog:

The Microchip IPC mailbox driver and IPC remoteproc drivers are
compatible with the Mi-V Inter-hart Communication (IHC) IP available
in the Libero catalog. Update the corresponding device trees to reflect
this change.

The new Mi-V IHC IP core is not backwards compatible with the Mi-V IHC
subsystem available in releases prior to v2025.07.

Link: https://openbeagle.org/beaglev-fire/gateware/-/commit/3758018476c166078e957e6f59662479e5554979
Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>