ARM: dts: microchip: sama7d65_curiosity: Add aliases for the SDMMC IPs
The SAMA7D65-Curiosity board has a SD Card slot on SDMMC1 and a SDIO M.2
KEY-E module on SDMMC2.
Add aliases for the enabled mmc devices, so that device numbering doesn't
vary depending on probe order.

Before the change, the 'root' argument on the command line could point to
an unintended boot device.

Signed-off-by: Aubin Constans <aubin.constans@microchip.com>
Acked-by: Nicolas Ferre <nicolas.ferre@microchip.com>